Update types, utils, constants
