+++ src/types/baidu.ts
@@ -1,0 +1,25 @@
+/**
+ * Baidu Options
+ */
+export interface BaiduOptions {
+  /**
+   * URL
+   *
+   * @type {string}
+   */
+  url: string;
+
+  /**
+   * Access Key
+   *
+   * @type {string}
+   */
+  key: string;
+
+  /**
+   * Callback Function
+   *
+   * @type {string}
+   */
+  callback: string;
+}
