+++ src/types/editor.ts
@@ -1,0 +1,20 @@
+import { Format } from '../constants';
+
+/**
+ * Editor Options
+ */
+export interface EditorOptions {
+  /**
+   * Height
+   *
+   * @type {number}
+   */
+  height: number;
+
+  /**
+   * Format
+   *
+   * @type {Format}
+   */
+  format: Format;
+}
