+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './baidu';
+export * from './editor';
+export * from './panel';
